back/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
SQLAlchemy==2.0.6
Werkzeug==2.2.3
python-dotenv==1.0.0
uuid==1.30
requests==2.28.2
python-binance==1.0.17
pytest==7.3.1
gunicorn==20.1.0
schedule==1.1.0
cryptography==39.0.2
sqlalchemy-utils==0.41.1
loguru==0.7.0 
PyJWT==2.6.0 

# 开发依赖
watchdog==3.0.0
flask-debugtoolbar==0.13.1 